+++ convey/__main__.py
@@ -50,19 +50,25 @@ def main():
         mod.post_mortem(tb)
 
 
+class WebServer:
+    source_parser = None
+
+
 def application(env, start_response):
     """ WSGI launcher. You may expose installed convey as a web service.
         Launch: uwsgi --http :9090 --wsgi-file wsgi.py
         Access: http://localhost:9090/?q=1.2.3.4
     """
-    from convey.config import Config
-    from convey.sourceParser import SourceParser
-    Config.init()
+    if not WebServer.source_parser:
+        from convey.config import Config
+        from convey.sourceParser import SourceParser
+        Config.init()
+        WebServer.source_parser = SourceParser(prepare=False)
 
     headers = [('Access-Control-Allow-Origin', '*')]
     t = env["QUERY_STRING"].split("q=")  # XX sanitize?
     if len(t) == 2:
-        response = SourceParser(stdin=[t[1]], prepare=False).check_single_cell()
+        response = WebServer.source_parser.set_stdin([t[1]]).check_single_cell()
         headers.append(('Content-Type', 'application/json'))
         status = '200 OK'
     else: