+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - new quote_char flag
 - new header / no-header flags
 - CIDR with host bits set ( = invalid network) translates to IP when asking whois (RIR would return 'invalid search key')
+- CIDR translated to an IP before asking whois
 
 ## 1.1 (2019-05-13)
 - fix: do not reprocess file if moved to another location